Move a few classes/enums into separate files.
