+++ osu.Game.Modes.Osu/Objects/Drawables/DrawableOsuHitObject.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
 using System.ComponentModel;
+using osu.Game.Modes.Judgements;
 using osu.Game.Modes.Objects.Drawables;
 
 namespace osu.Game.Modes.Osu.Objects.Drawables
