+++ osu.Game/Modes/Judgements/JudgementInfo.cs
@@ -1,0 +1,11 @@
+﻿using osu.Game.Modes.Objects.Drawables;
+
+namespace osu.Game.Modes.Judgements
+{
+    public class JudgementInfo
+    {
+        public ulong? ComboAtHit;
+        public HitResult? Result;
+        public double TimeOffset;
+    }
+}
