+++ osu.Game/Modes/Objects/Drawables/ArmedState.cs
@@ -1,0 +1,9 @@
+namespace osu.Game.Modes.Objects.Drawables
+{
+    public enum ArmedState
+    {
+        Idle,
+        Hit,
+        Miss
+    }
+}
