+++ osu.Game/Modes/Objects/Drawables/HitResult.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel;
+
+namespace osu.Game.Modes.Objects.Drawables
+{
+    public enum HitResult
+    {
+        [Description(@"Miss")]
+        Miss,
+        [Description(@"Hit")]
+        Hit,
+    }
+}
